refactor: - made the amount column a decimal instead of a float - Added Account Name to transactions table - Only put the fields that were needed in the TransactionDTO - Added extension for to transform Transaction object into TransactionDTO object

+++ Server/Data/Models/MonthlySummary.cs
@@ -4,7 +4,7 @@ public record MonthlySummary
 {
     public required string Month { get; set; }
     
-    public double Income { get; set; }
+    public decimal Income { get; set; }
     
-    public double Expenditure { get; set; }
+    public decimal Expenditure { get; set; }
 }
